refactor(platform): extract app domain service worker
